requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.19.0
pyyaml>=6.0
accelerate>=0.30.0
